+++ App/src/analytics/AuditEventName.ts
@@ -1,0 +1,9 @@
+export enum AuditEventName {
+    CreateSMS = "Create SMS",
+    OpenUrl = "Open URL",
+    MakeCall = "Make Call",
+    SendEMail = "Send EMail",
+
+    SaveFavorites = "Save Favorites",
+    BackgroundSync = "Background Sync",
+}