+++ App/src/analytics/AuditScreenName.ts
@@ -1,0 +1,21 @@
+export enum AuditScreenName {
+    SignIn          = "Auth/SignIn",
+    ConfirmSignIn   = "Auth/Confirm",
+
+    Member          = "Members/Member",
+    Club            = "Structure/Club",
+    Web             = "Web",
+
+    Associations    = "Structure/Associations",
+    Areas           = "Structure/Areas",
+    Clubs           = "Structure/Clubs",
+
+    MemberList      = "Members/List",
+    FilterMember    = "Members/Filter",
+    MemberSearch    = "Members/Search",
+
+    Settings        = "Settings",
+    Legal           = "Settings/Legal",
+    ShowMD          = "Settings/Show MD",
+    ShowExternal    = "Settings/Show External",
+}