+++ App/src/components/Club/RoleAccordionSection.tsx
@@ -1,11 +1,12 @@
 import gql from 'graphql-tag';
 import React from 'react';
 import { Query } from 'react-apollo';
-import { withCacheInvalidation } from '../../helper/cache/withCacheInvalidation';
+import { isRecordValid } from '../../helper/cache/withCacheInvalidation';
 import { sortGroupRoles } from '../../helper/sortRoles';
 import { Accordion } from '../Accordion';
-import { InlineLoading } from '../Loading';
+import { Placeholder } from '../Placeholder/Placeholder';
 import { RoleCard } from './RoleCard';
+import { RolesPlaceholder } from "./RolesPlaceholder";
 
 type Props = {
     group: string,
@@ -30,6 +31,7 @@ const MemberFragment = gql`
 const RolesQuery = gql`
     query Roles($id: String!) {
         Club(id: $id) {
+            LastSync @client
             id
 
             board {
@@ -55,24 +57,36 @@ class RoleAccordionSectionBase extends React.Component<Props, State> {
     render() {
         return (
             <Accordion title={this.props.group} expanded={this.props.expanded}>
-                <Query query={RolesQuery} variables={{id: this.props.club}} fetchPolicy={this.props.fetchPolicy}>
-                    {({ loading, data, /*error, data, */refetch }) => {
-                        if (!loading && data != null) {
-                            const grouped = sortGroupRoles(data.Club[this.props.groupDetails]);
-                            if (grouped == null) return null;
-                            const len = grouped.length;
-
-                            return grouped.map((r, i) => (
-                                <RoleCard
-                                    key={r.member + r.role}
-                                    member={r.member}
-                                    role={r.role}
-                                    separator={i !== len - 1}
-                                />
-                            ));
+                <Query query={RolesQuery} variables={{ id: this.props.club }}>
+                    {({ loading, data, refetch }) => {
+                        if (data && data.Club != null) {
+                            if (!isRecordValid("club", data.Club.LastSync)) {
+                                setTimeout(() => refetch({ id: this.props.club }));
+                            }
                         }
 
-                        return <InlineLoading />
+                        let grouped: any = null;
+                        let len = 0;
+
+                        if (!loading && data != null) {
+                            grouped = sortGroupRoles(data.Club[this.props.groupDetails]);
+                            len = grouped ? grouped.length : 0;
+                        }
+
+                        return (
+                            <Placeholder ready={!loading && data != null} previewComponent={<RolesPlaceholder count={3} />}>
+                                {
+                                    grouped && grouped.map((r, i) => (
+                                        <RoleCard
+                                            key={r.member + r.role}
+                                            member={r.member}
+                                            role={r.role}
+                                            separator={i !== len - 1}
+                                        />
+                                    ))
+                                }
+                            </Placeholder>
+                        );
                     }}
                 </Query>
             </Accordion>
@@ -80,4 +94,4 @@ class RoleAccordionSectionBase extends React.Component<Props, State> {
     }
 }
 
-export const RoleAccordionSection = withCacheInvalidation("clubs", RoleAccordionSectionBase);
+export const RoleAccordionSection = RoleAccordionSectionBase;