+++ App/src/screens/Filter/Queries.tsx
@@ -1,4 +1,5 @@
 import gql from 'graphql-tag';
+
 export const AreasQuery = gql`
     query AreasQuery {
         Areas {